Update code move file

+++ folders/move_files_to_folder.py
@@ -16,12 +16,16 @@ def convert_string_to_date(date_string):
 def move_files(path):
     with os.scandir(path) as itr:
         for entry in itr:
+            if not entry.is_file():
+                continue
+
             file_name = entry.name
             pattern = r"\d{4}-\d{2}-\d{2}"
             match = re.search(pattern, file_name)
             if match:
                 date_time = match.group()
                 path_new = f"{os.path.join(path, date_time)}"
+                
                 if os.path.isdir(path_new) is False:
                     os.mkdir(path=path_new)
                 shutil.move(entry.path, f"{os.path.join(path, date_time)}")
